requirements.txt
autopep8==1.5.7
beautifulsoup4==4.10.0
certifi==2021.5.30
charset-normalizer==2.0.6
click==8.0.1
Flask==2.0.1
flickrapi==2.4.0
idna==3.2
importlib-metadata==4.8.1
itsdangerous==2.0.1
Jinja2==3.0.1
loguru==0.5.3
lxml==4.6.3
MarkupSafe==2.0.1
oauthlib==3.1.1
pycodestyle==2.7.0
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
six==1.16.0
soupsieve==2.2.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.1
zipp==3.5.0
